auto-run cleanup after test context is finished

+++ solidblocks-cli/src/main/kotlin/de/solidblocks/cli/hetzner/asg/HetznerAsg.kt
@@ -387,7 +387,7 @@ class HetznerAsg(hcloudToken: String) {
                         serverTypeRef.reference,
                         imageRef.reference,
                         placementGroupRef?.reference,
-                        sshKeyRefs.map { it.reference },
+                        sshKeyRefs.map { it.reference.toLong() },
                         networks.map { it.id.toString() } + networkRefs.map { it.reference },
                         firewallRefs.map { it.reference },
                         userData,
